Stop treating linked libraries as in-project files
Currently files from linked libraries are synced just like normal
project files. This causes the same library file to exist multiple times
on the local disk (in every project it is linked to). If one of the
local versions is updated, this update is currently not propagated to
the other local versions of the file.

To fix this the CLI will stop treating files from linked libraries as
in-project files. This means that only the original versions (the
version of the file in its library project in the 'Library' directory)
will be synced. By working this way, every library file only exists once
on the local disk (in its project in the 'Library' directory).

If a user tries to push, pull or delete a library file in a non-library
project a warning will be shown letting him know of the above. Users
can safely remove local versions of library files in non-library
projects.

Because this change makes it impossible to know which project has access
to which files in a QuantConnect CLI project (there is no way to see
which libraries are linked to a project anymore) a projects:info command
will be added soon which will display this information.